pyproject.toml
[project]
name = "niuma"
version = "0.1.0"
description = "A cognitive multi-agent AI system"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Niuma Team" }
]
keywords = ["ai", "agent", "llm", "multi-agent", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # 核心框架
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "anyio>=4.0",

    # CLI
    "typer>=0.9",
    "rich>=13.0",

    # LLM
    "openai>=1.0",
    "anthropic>=0.20",
    "langchain>=0.1",

    # 记忆存储
    "chromadb>=0.4",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",

    # MCP
    "mcp>=1.0",

    # 工具
    "gitpython>=3.1",
    "aiofiles>=23.0",
    "httpx>=0.27",

    # 配置
    "pyyaml>=6.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
]

[project.scripts]
niuma = "niuma.cli.main:app"

[project.urls]
Homepage = "https://github.com/niuma-ai/niuma"
Documentation = "https://docs.niuma.ai"
Repository = "https://github.com/niuma-ai/niuma"
Issues = "https://github.com/niuma-ai/niuma/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["niuma"]

# ============================================
# Ruff 配置 - 代码格式化和 Lint
# ============================================
[tool.ruff]
target-version = "py311"  # Python 3.11+
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E", "W",
    # isort
    "I",
    # Pyupgrade
    "UP",
    # Flake8-bugbear
    "B",
    # Flake8-simplify
    "SIM",
    # Flake8-comprehensions
    "C4",
    # Flake8-unused-arguments
    "ARG",
    # Pydocstyle
    "D",
    # McCabe complexity
    "C90",
]
ignore = [
    # 文档字符串风格（使用 Google 风格）
    "D203",  # one-blank-line-before-class
    "D213",  # multi-line-summary-second-line
    # 其他
    "D104",  # Missing docstring in public package
    "D107",  # Missing docstring in __init__
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
# 测试文件放宽规则
"tests/**/*" = ["D", "ARG"]
"__init__.py" = ["D104"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ============================================
# MyPy 配置 - 类型检查
# ============================================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# 第三方库（无类型支持）
[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "git",
    "mcp.*",
]
ignore_missing_imports = true

# ============================================
# Pytest 配置 - 测试
# ============================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"

# ============================================
# Coverage 配置 - 代码覆盖率
# ============================================
[tool.coverage.run]
source = ["niuma"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
fail_under = 80

[tool.coverage.html]
directory = "htmlcov"

# ============================================
# 开发依赖组 (UV 2.0+)
# ============================================
[dependency-groups]
dev = [
    # 测试
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-anyio>=0.0",

    # Lint & Format
    "ruff>=0.4",
    "mypy>=1.9",

    # 开发工具
    "pre-commit>=3.7",
    "ipython>=8.0",
]

# ============================================
# UV 特定配置
# ============================================
[tool.uv]
